bmad/requirements.txt
# BMAD Requirements
# Core dependencies for Business Model As Documents

# Core system
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2

# Data processing and analysis
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Database and caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.5

# BMAD-specific dependencies
# For dynamic story selection
# (None currently needed beyond the above)

# Optional dependencies for enhanced features
# Uncomment if needed:
# transformers==4.35.2
# spacy==3.7.2
# jupyter==1.0.0
# plotly==5.17.0

# Development dependencies
# For integration testing and development
pytest-mock==3.11.1
faker==19.3.1
requests==2.31.0

# Async support
aiohttp==3.9.1
websockets==12.0

# Monitoring and logging
structlog==23.1.0
prometheus-client==0.17.1

# Configuration management
pydantic-settings==2.0.3
dynaconf==3.1.9
